requirements.txt
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
clickclick==20.10.2
connexion==2.9.0
cryptography==3.4.7
Flask==1.1.4
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==3.2
inflection==0.5.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==3.2.0
MarkupSafe==2.0.1
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
pycparser==2.20
PyJWT==2.1.0
pyrsistent==0.18.0
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.22
swagger-ui-bundle==0.0.8
urllib3==1.26.6
Werkzeug==1.0.1
